refactor(ios): split runner startup transport
